Make get_endpoint a generic reusable function.
This patch refactors the get_endpoint function to make it more generic
and top level so that it can be more easily accessed and tested.

Adds service_type, endpoint_region, and endpoint_type options to
the get_endpoint function as well.

Added a new test suite specifically for get_endpoint.

The motivation for this change is that we'll need to be able to
parse the 'object-store' endpoint for each tenant in order to
enable multi-tenant storage in the swift storage backend.

Partially implements blueprint: swift-tenant-specific-storage

Change-Id: Icf773a1eb057fc9d65cf54890afa4638f2aee0b4