Provide stores access to the request context.
Adds a set_context function for the Store class which can be used
to provide an instantiated store access to the RequestContext.

Includes the following changes:

* Updates common store helper methods to call set_context on each
  store as it is retrieved from the list of know stores.

* Implement custom set_context call on the swift store.

* Updates Glance scrubber to create a real context from credentials.

The motivation for this change is that we will need access to the
service catalog and token for the Swift multi-tenant backend.

Partially implements blueprint: swift-tenant-specific-storage

Change-Id: I3def2b7c4085a36dd6c0961e762783ebaf7ffa74